+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##
 ## Added 
 - ability to add multiple rooms/files to a page
+- widget setting (view mode 'editor/embedded')
 
 ##
 ## Changed